Update and rename OHMs-ArduinoMicro v0.54 to OHMs-ArduinoMicro v0.66
Lots of hours of testing and playing the instrument later......
"Primitive Protocol" has been abandoned for now. Instead I have focused my efforts on a stand-alone synth for the OHMs based on Pure Data.  the "Interval" MIDI-message section has been cleaned up quite a lot, mainly because of the troubleshooting I've had to do to figure out Pure Data.
This version (v0.66) works with both the standard (breath-controller) synths in my ancient version of Reason, and with my new homemade synth in PD. Monophonic and polyphonic playing both work fine, as well as breath and bend.