Update and rename OHMs-ArduinoMicro v0.67 to OHMs-ArduinoMicro v0.70
Total rewrite of the Midi Message section. Much shorter, and now voice specific, so "currentNote" goes to channel 1, and "intervalNote" to channel 2. 
This new output seems like a better way of dealing with this kind of two voice polyphony. I can now easily put a different synth on the second voice, or pass it through different filters than the principal voice, if I want.
I'm making a synth for the OHMs in Pure Data, which lets me take full advantage of this.

v0.70 also has a couple of new extra fingerings in the "pickNote()" function:
Bb can now also be fingered as a Bnatural + F or F# key (or both)
Cnatural can be fingered as a Bnatural + Bb side key.