OHMs-Teensy3.1 v0.75
Total rewrite of the code for the OHMs.
Electronics have changed: The OHMs is now built on a Teensy 3.1 with custom breakout board.
Key-logic has changed. Also, there are now up to 22 keys instead of 16.
There is a new sensor: The MPU-9250 inertial measurement unit. I am currently reading only its gyroscope.
There are now 3 separate calibration buttons, "Breath", "bend" and "IMU", for on the fly calibration without the need for a menu system.
The OHMs currently has a "Microview" display connected, but there is no menu implemented yet.

There are still many things missing here, but the speed and functionality are already much better than the old code running on the old electronics. The calibration buttons mean that the menu will now mainly be needed for troubleshooting and memory management. As far as playing, the OHMs 0.75 is more than ready, in fact I'm playing it in concert tonight :-)