Update and rename OHMs-Teensy3.1 v0.82 to OHMs-Teensy3.1 v0.86e_exp
Implemented high resolution analog sensor readings using the ADC library. Breath and pitchbend are now read at 12 bits (0-4095), and are filtered and mapped to 0-127 for a super smooth output. (TODO: ANALOG OUTPUT ON PIN 14)
Implemented the  MPU-9250 IMU (Inertial Measurement Unit), plus a 128x64 OLED display. Both run on the same i2c bus, pins 18-19.
This is my second shot at implementing the IMU on the OHMs, and this time it's actually working :)
The IMU outputs cc1 and cc4 right now, but this will all be selectable in the menu. Furthermore, one axis is used as a multiplier for the other, so "tilting" side to side starts having a more noticeable effect as I raise the OHMs to a more horizontal angle. After days of calibrating / playing around, the output is now smooth enough to use for music.
IMU-calibration is hard. The "compass-angle" is not implemented as sensor input yet, simply because I still have too much drift in the magnetometer data.
The display is only used for IMU-calibration for now, but I'm quite happy with it, so the new menu system with the rest of the calibration options etc. will be based on it.


Side note 2: The IMU sensor-fusion math is also hard, and it has measurably slowed down the execution of the main loop of the OHMs sketch. Maximum loop time is now 5ms so it's still plenty good enough for music. I say "measurably", because the loop time has been fluctuating with each version since I started measuring it, some versions much slower than this. I will look for ways to improve this, once I understand the IMU a little better. 
When the display is on (for calibration / menu access), the max loop time is around 100ms. For now the sequence is: 30ms of sensor reading and sending MIDI, followed by a display update. This gives a noticeable lag to everything, but it works fine for checking if your calibration is working. Exit the menu, and you're straight back in the fast lane.