Update and rename OHMs-Teensy3.1 v0.86e_exp to OHMs-Teensy3.1 v0.89c_exp
Complete rewrite of much of the code + reorganization of all of it.
Various functions are now spread out over many header (.h) files.
The new "readAnalogSensor_COMPLEX" function reads all analog sensors (up to 10). It also applies Median and vector filters, as well as mapping, input and output slope correction. Parameters for each sensor is held in a struct, which is updated with calibration data, as well as changes in the menu (still to be implemented)
Menu is in beginning stages, with only the mainMenu and subMenus functional.